Refactored some of the shared CLI code in 'rct'
