+++ bin/rct
@@ -33,18 +33,16 @@ configure_i18n()
 import gettext
 _ = gettext.gettext
 
-from rct.cli import CLI
-from rct.commands import CatCertCommand
+from rct.cli import RctCLI
 
 # Make sure that we have root access
 if os.getuid() != 0:
-    sys.stderr.write('Error: this command requires root access to execute\n')
+    sys.stderr.write(_('Error: this command requires root access to execute') + '\n')
     sys.exit(8)
 
 
 def main():
-    rct_commands = [CatCertCommand]
-    return CLI(commands=rct_commands).main()
+    return RctCLI().main()
 
 
 if __name__ == '__main__':
@@ -55,5 +53,5 @@ if __name__ == '__main__':
         #re-raise, bypassing handle_exception
         raise e
     except KeyboardInterrupt:
-        sys.stderr.write("\nUser interrupted process.")
+        sys.stderr.write("\n" + _("User interrupted process."))
         sys.exit(0)
